// src/types/reducer-types.ts
import { CartItem, ShippingInfo } from "./types";
// src/types/reducer-types.ts

import { User } from "./types"; // Adjust this path if needed

export interface UserReducerInitialState {
  user: User | null;
  loading: boolean;
}

export interface CartReducerInitialState {
  loading: boolean;
  cartItems: CartItem[];
  subtotal: number;
  // tax: number;
  shippingCharges: number;
  discount: number;
  total: number;
  shippingInfo: ShippingInfo;
}